+++ nodes/device-read.js
@@ -2,7 +2,10 @@ module.exports = function(RED) {
 
     "use strict";
 
+    const Utils = require('../lib/utils/utils');
+
     function DeviceReadNode(config) {
+
         RED.nodes.createNode(this, config);
 
         // get node
@@ -15,7 +18,10 @@ module.exports = function(RED) {
         node.on('input', function(msg, send, done) {
 
             let device = config.device || msg.device;
+            device = Utils.mustacheRender(device, msg);
+
             let resource = config.resource || msg.resource;
+            resource = Utils.mustacheRender(resource, msg);
 
             const method = 'POST';
             const url = `${server.config.ssl ? "https://" : "http://"}${server.config.host}/v3/users/${server.config.username}/devices/${device}/resources/${resource}`;