+++ nodes/endpoint-call.js
@@ -2,6 +2,8 @@ module.exports = function(RED) {
 
     "use strict";
 
+    const Utils = require('../lib/utils/utils');
+
     function EndpointCallNode(config) {
         RED.nodes.createNode(this, config);
 
@@ -15,12 +17,24 @@ module.exports = function(RED) {
         node.on("input",function(msg, send, done) {
 
             let endpoint = config.endpoint || msg.endpoint;
+            endpoint = Utils.mustacheRender(endpoint, msg);
+
+            let data = Utils.transformValue(config.value);
+            if ( data === null ) {
+                if ( typeof msg.payload !== 'undefined' ) {
+                    data = Utils.transformValue(msg.payload);
+                }
+                else if ( typeof msg.value !== 'undefined' ) {
+                    data = Utils.transformValue(msg.payload);
+                }
+            }
+            data = Utils.mustacheRender(data, msg);
 
             const method = 'POST';
             const url = `${server.config.ssl ? "https://" : "http://"}${server.config.host}/v1/users/${server.config.username}/endpoints/${endpoint}/call`;
 
             if (typeof server.request === "function") {
-              server.request(node, url, method, msg.payload)
+              server.request(node, url, method, data)
               .then(res => {
 
                   // Throw if response fails